feat: Add fuctionality for filtering books
