+++ index.html
@@ -37,7 +37,6 @@
             aria-label="Search"
           />
         </div>
-
         <ul class="navbar-nav ">
           <li class="nav-item">
             <button
